refactored participant feature

+++ src/Features/Participant/Controllers.ts
@@ -1,18 +1,40 @@
-import prisma from "../../Utils/Prisma";
-import { Static } from "@sinclair/typebox";
 import { FastifyRequest, FastifyReply } from "fastify";
-import { eventUserIdParams } from "./Contracts";
 
-// Remove a user from an event
-export const removeUserEventController = async (req: FastifyRequest<{ Params: Static<typeof eventUserIdParams> }>, res: FastifyReply) => {
-  try {
-    const userID = await prisma.eventParticipant.delete({
-      where: { UserID_EventID: { EventID: req.params.eventID, UserID: req.params.userID } },
-    });
-    const user = await prisma.user.findUnique({ where: { ID: userID.UserID } });
+import prisma from "../../Utils/Prisma";
+import * as Contracts from "./Contracts";
 
-    return res.status(200).send(user);
-  } catch (error) {
-    return res.status(500).send(error);
-  }
+// Remove a Participant from an Event
+export const removeParticipantByOwnerController = async (
+	req: FastifyRequest<{
+		Params: typeof Contracts.DeleteParticipantByOwnerSchema.params.static;
+		Body: typeof Contracts.DeleteParticipantByOwnerSchema.body.static;
+	}>,
+	res: FastifyReply,
+) => {
+	try {
+		// Verify if the event exists and check for permission
+		const event = await prisma.event.findUnique({ where: { ID: req.params.EventID } });
+		if (!event) throw new Error("Event does not exist");
+		if (event.UserID !== req.user.ID) throw new Error("You do not have permmission");
+
+		// The user must exist
+		const user = await prisma.user.findUnique({ where: { ID: req.body.UserID } });
+		if (!user) throw new Error("User does not exist");
+
+		// The user must have already asked for permission
+		const participant = await prisma.eventParticipant.findUnique({
+			where: { UserID_EventID: { UserID: user.ID, EventID: req.params.EventID } },
+		});
+		if (!participant) throw new Error("User is not a participant");
+
+		// We delete the participant and update the current user on the event
+		await prisma.eventParticipant.delete({
+			where: { UserID_EventID: { EventID: req.params.EventID, UserID: req.body.UserID } },
+		});
+		await prisma.event.update({ where: { ID: event.ID }, data: { CurrentUsers: event.CurrentUsers - 1 } });
+
+		return res.status(200).send({});
+	} catch (error) {
+		return res.status(500).send(error);
+	}
 };
